improve flexibility of transaction logic -> move to hooks
