+++ explorer/src/types/transaction.ts
@@ -1,4 +1,5 @@
-import type { Chain } from 'constants/chains'
+import type { SubmittableModuleExtrinsics } from '@autonomys/auto-utils'
+import type { Chain, NetworkSource } from 'constants/chains'
 import type { TransactionStatus } from 'constants/transaction'
 import type { WalletAccountWithType } from 'types/wallet'
 
@@ -21,4 +22,36 @@ export interface Transaction {
 export interface TransactionWithMetadata extends Transaction {
   submittedAtLocalTimestamp: Date
   finalizedAtLocalTimestamp?: Date
+}
+
+export interface OptionalTxFormValues {
+  nonce?: number
+}
+
+export type ExtrinsicModule = SubmittableModuleExtrinsics<'promise'>
+export type ExtrinsicsList = { [key: string]: ExtrinsicModule }
+export type ExtrinsicsMethodFields = { name: string; type: string; typeName: string }
+export type ExtrinsicsMethodArgs = ExtrinsicsMethodFields & { docs: string[] }
+
+export type ExtrinsicsMethod = {
+  name: string
+  docs: string[]
+  fields: ExtrinsicsMethodFields[]
+  args: ExtrinsicsMethodArgs[]
+}
+export interface CustomExtrinsicFormValues {
+  [key: string]: string
+}
+export interface MessageFormValues extends OptionalTxFormValues {
+  message: string
+}
+
+export interface SendTokenFormValues extends OptionalTxFormValues {
+  sender: string
+  sourceNetwork: NetworkSource
+  sourceDomainId: string
+  receiver: string
+  destinationNetwork: NetworkSource
+  destinationDomainId: string
+  amount: number
 }
